requirements.txt
certifi==2019.3.9
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==19.9.0
numpy==1.16.2
pandas==0.24.2
psycopg2==2.8
python-dateutil==2.8.0
pytz==2018.9
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
whitenoise==4.1.2
wincertstore==0.2
